requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# main.py: 1
# main2.py: 3
Flask == 1.0.2

# forms.py: 1
Flask_WTF == 0.14.2

# single_image_prediction.py: 5
Pillow == 5.1.0

# forms.py: 3
WTForms == 2.2.1

# main2.py: 4
Werkzeug == 0.14.1

# single_image_prediction.py: 3
numpy == 1.14.3

# single_image_prediction.py: 2
pandas == 0.23.0

# main.py: 6
# single_image_prediction.py: 4
torch == 0.4

# main.py: 7
# single_image_prediction.py: 7
torchvision == 0.2.1
